% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{create_abundance_table}
\alias{create_abundance_table}
\title{Creates an abundance matrix table}
\usage{
create_abundance_table(filtered_results)
}
\arguments{
\item{filtered_results}{the filtered results produced by dada2}
}
\value{
an abundance table
}
\description{
`create_abundance_table` is used to create an abundance matrix table
which will be used for heatmap generation
}
