% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{create_heatmap}
\alias{create_heatmap}
\title{Creates heatmap}
\usage{
create_heatmap(table)
}
\arguments{
\item{table}{abundance table generated from `create_abundance_table`}
}
\value{
None. Generates a heatmap
}
\description{
`create_heatmap` takes in an abundance table to generate a heatmap
to illustrate
}
