% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{preprocess_16s_data}
\alias{preprocess_16s_data}
\title{Preprocess 16S Data}
\usage{
preprocess_16s_data(inputs, outputs, verbose = FALSE, multithread = FALSE)
}
\arguments{
\item{verbose}{set to true for explicit updates in processing}

\item{multithread}{set to true if you wish to utilize multiple CPU cores}

\item{input_fastq_files}{directory containing desired input files}

\item{output_files}{directory where output files will go}
}
\value{
a dada output file
}
\description{
`preprocess_16s_data` preprocesses 16S rRNA single or paired-end reads
that are in FASTQ format.
}
