% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_heatmap.R, R/developing.R
\name{create_heatmap}
\alias{create_heatmap}
\title{Generates a heatmap to illustrate the abundance of organisms across sample microbiomes.}
\usage{
create_heatmap(
  table,
  proportion = 1,
  low_col = "#F2E7C9",
  high_col = "#801A86"
)

create_heatmap(
  table,
  proportion = 1,
  low_col = "#F2E7C9",
  high_col = "#801A86"
)
}
\arguments{
\item{table}{abundance table generated from `create_abundance_table`}

\item{proportion}{proportion of table between 0 and 1 to visualize. If 1, will visualize all columns, if 0.5 will visualize the first half of the columns of the table.}

\item{low_col}{a colour to indicate low values on the heatmap}

\item{high_col}{a colour to indicate high values on the heatmap}
}
\value{
A ggplot2 heatmap object

A ggplot2 heatmap object
}
\description{
`create_heatmap` takes in an abundance matrix to generate a heatmap
to illustrate abundance of 16S sequence variants across samples.

`create_heatmap` takes in an abundance matrix to generate a heatmap
to illustrate abundance of 16S sequence variants across samples.
}
\examples{
\dontrun{
heatmap <- create_heatmap(table, proportion = 0.025)
}

plots_dir = "data/graphics"
\dontrun{
heatmap <- create_heatmap(table, plots_dir, proportion = 0.025)
}

}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.

Wickham H, Vaughan D, Girlich M (2023). tidyr: Tidy Messy Data.
R package version 1.3.0, <https://CRAN.R-project.org/package=tidyr>.

Müller K, Wickham H (2023). tibble: Simple Data Frames.
R package version 3.2.1, <https://CRAN.R-project.org/package=tibble>.

H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.

Wickham H, Vaughan D, Girlich M (2023). tidyr: Tidy Messy Data.
R package version 1.3.0, <https://CRAN.R-project.org/package=tidyr>.

Müller K, Wickham H (2023). tibble: Simple Data Frames.
R package version 3.2.1, <https://CRAN.R-project.org/package=tibble>.
}
\author{
{Koji Wong, \email{koji.wong@mail.utoronto.ca}}
}
