% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{preprocess_16s_data}
\alias{preprocess_16s_data}
\title{Preprocess 16S Sequence Data using Dada2 Package}
\usage{
preprocess_16s_data(
  input_dir,
  output_dir,
  verbose = FALSE,
  multithread = FALSE
)
}
\arguments{
\item{input_dir}{directory containing desired input files. Must be in fastq.gz, fastq.bz2, or .fastq file}

\item{output_dir}{target directory where preprocessed data will be saved. This is the same directory that will be used to learn error rates.}

\item{verbose}{set to true for explicit updates during preprocessing}

\item{multithread}{set to true if you wish to utilize multiple CPU cores for extra speed.}
}
\value{
a dada output file
}
\description{
`preprocess_16s_data` preprocesses 16S rRNA single or paired-end reads
that are in FASTQ format. This function calls the `filterAndTrim` function from
the dada2 package which filters and trims fastq file inputs.
}
\examples{
# Example:
# Directory containing raw single-end 16S reads
input <- system.file("inst/extdata/sample_raw_16S_data", package = "CompHeatmaps")
output <- system.file("/inst/extdata/filtered_reads", package = "CompHeatmaps")
\dontrun{
library("CompHeatmaps")
# Set verbose to TRUE for updates on progress in terminal and multithread to TRUE for faster computation if you have multiple cores.
CompHeatmaps::preprocess_16s_data(input_dir = input, output_dir = output, verbose = TRUE, multithread = TRUE)
length(result) # 4
class(result) # list
class(result[[1]]) # dada2
}

}
\author{
Koji Wong
}
