% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCompHeatmaps.R
\name{runCompHeatmaps}
\alias{runCompHeatmaps}
\title{Launch Shiny App for CompHeatmaps}
\usage{
runCompHeatmaps()
}
\value{
No return value, opens up a Shiny page.
}
\description{
A function that launches the Shiny app for CompHeatmaps.
The purpose of this app is to allow users to interact and create heatmaps on the fly from raw data.
app works. The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
CompHeatmaps::runCompHeatmaps()
}

}
